mips: fix configure issue with specific paths
If configure is invoked with e.g.

CFLAGS='--sysroot=/workspace/linux-mips64/sysroot -mabi=64'

flags FLAG_M32 and FLAG_M64 are not set, because the grep search in
configure.ac mismatches the mips keyword in the path.

Update the grep patterns to anchor on non-word boundaries.

This fixes KDE #412344.

Patch by Chris Packham.