Bug 418435 - s390x: Avoid extra value dependency in CLC implementation
The test memcheck/tests/memcmp currently fails on s390x because it yields
the expected "conditional jump or move depends on uninitialised value(s)"
message twice instead of just once.

This is caused by the handling of the s390x instruction CLC, see
s390_irgen_CLC_EX().  When comparing two bytes from the two input strings,
the implementation uses the comparison result for a conditional branch to
the next instruction.  But if no further bytes need to be compared, the
comparison result is also used for generating the resulting condition
code.

There are two cases: Either the inputs are equal; then the resulting
condition code is zero.  This is what happens in the memcmp test case.  Or
the inputs are different; then the resulting condition code is 1 or 2 if
the first or second operand is greater, respectively.

At least in the first case it is easy to avoid the additional dependency,
by clearing the condition code explicitly.  Just do this.