s390x: Don't emit "vector or with complement" on z13
The z/Architecture instruction "vector or with complement" (VOC) can be
used as an optimization to combine "vector or" with "vector nor".  This is
exploited in Valgrind since commit 6c1cb1a0128b00858b973e.  However, VOC
requires the vector-enhancements facility 1, which is not installed on a
z13 CPU.  Thus Valgrind can now run into SIGILL on z13 when trying to
execute vector string instructions.

Fix this by suppressing the VOC optimization unless the
vector-enhancements facility 1 is recognized on the host.