Remove deprecated regression tests for mftgpr and mffgpr.
The mftgpr and mffgpr instructions are deprecated.  Added comments in
VEX/priv/guest_ppc_toIR.c for the instructions stating the instructions
are deprecated.  Valgrind support can be removed if the opcodes get reused
in the future.  For now, leaving the functional support in Valgrind
for the instructions.

Removed the regression test power6_mf_gpr.c, expect files and vgtest file
from none/tests/ppc64.